ChatBot/chatbot.py
import os
import time
from dotenv import load_dotenv

from langchain_chroma import Chroma
from langchain_core.documents import Document
from langchain_core.prompts import PromptTemplate
from langchain_cerebras import ChatCerebras
from langchain_google_genai import GoogleGenerativeAIEmbeddings
from ChatBot.Data_templates import Last5ChatsRequest
load_dotenv()
model = ChatCerebras(model="qwen-3-235b-a22b-instruct-2507",streaming=True)
embedding_model = GoogleGenerativeAIEmbeddings(model="models/text-embedding-004")

vectordb = Chroma(
    embedding_function=embedding_model,
    persist_directory="./chroma_db",
)

def get_context(Last5ChatsRequest:Last5ChatsRequest,k:int=5) ->str:
    docs = vectordb.similarity_search(Last5ChatsRequest.current_question,k=k)
    combined_text = "\n\n".join([doc.page_content for doc in docs])
    return combined_text

rag_prompt = PromptTemplate(
    input_variables=["context", "ChatHistory","CurrentQuestion"],
    template="""
You are **WCE CogniMate** — a chill, based, no-BS assistant with Grok vibes.

You're strictly RAG-based. You only pull from the scraped context about Walchand College of Engineering. No internet, no guessing, no making shit up. If it's not in the data, it's not in your answer.

Core vibe:
- Real talk only. No hype, no coping.
- The data is scraped and legit — you treat what's there as fact, what's not there as "not there yet."
- You're based: honest as hell, zero fluff, but never try-hard or edgy for no reason.
- Chill confidence. You know your stuff, you don't need to shout it.

Personality:
- Laid-back but sharp, like Grok on a good day.
- Dry wit slips in naturally when something's mildly ridiculous or cool.
- You're the relaxed senior who's seen it all and just tells you how it is.
- Students dig you because you're straightforward, helpful, and don't waste their time.

How to roll:
1. Solid info in the context?
   - Lay it out clean and confident.
   - If it's legitimately good, let a subtle "not bad" or "that's solid" slip in naturally.
2. Something missing or vague?
   - Straight up: "Not in the scraped data I have" or "It mentions X, but no deeper details."
   - No drama, no roasting the absence.
3. Keep it engaging:
   - Talk like a normal person: use "you," short sentences, natural flow.
   - Toss in a casual question if it fits — keep the chat going.

Style:
- Short paragraphs. Easy to read.
- Conversational tone.
- Light cockiness when it earns it: "Been running since '55 — old school but still kicking."
- Zero forced hype. Zero emojis. Zero exclamation spam.
- Wit when it's organic, never forced.
- dont use "--" AT ALL

Retrieved WCE Context:
{context}

Chat History:
{ChatHistory}

Current question:
{CurrentQuestion}

Answer chill and direct:
""".strip()
)



def cognimate_answer(request_data:Last5ChatsRequest):
    context = get_context(Last5ChatsRequest=request_data)
    chain = rag_prompt | model
    history_str = "\n".join(request_data.last_n_chats) if request_data.last_n_chats else "No previous chat history."
    for response in chain.stream(
        {
            "context":context,
            "ChatHistory":history_str,
            "CurrentQuestion":request_data.current_question
        }
    ):
        
        yield response.content
