requirements.txt
fastapi
uvicorn[standard]
chromadb


langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-chroma

langchain-cerebras

pydantic
python-dotenv

beautifulsoup4
requests

tiktoken
numpy
